Update defaultElements, add additional delegate, add disabled skin for DefaultUISkin
