+++ korge/src/commonMain/kotlin/com/soywiz/korge/ui/defaultImages.kt
@@ -1,0 +1,10 @@
+package com.soywiz.korge.ui
+
+import com.soywiz.korim.format.*
+import com.soywiz.korio.util.encoding.*
+
+val DEFAULT_UI_SKIN_IMG by lazy {
+    PNG.decode(
+        "iVBORw0KGgoAAAANSUhEUgAAAQAAAABACAYAAAD1Xam+AAAIv0lEQVR4nO2dW2wcVxnHf+td27vexOtN7CTOzbEdRSGWIGqKQhxzU81D1Rci1UkLaUEiLxgsLhJBIAQCXqoWeKipkSLxQImCiAMJQiVpFdQXN22joBbUJihdJw7KxqSR48v6srfZw8N6o9kzCx5Ed8zOd37SSPY3Z6Xvvzvz3++c+XYm0NfbTwUagM8ubweAzcuxWmIeuA28BZwF/gBkXb7Wl/pjsbho/VT4/Ds62yq9Voz+QAUDOAQ8C+yscoJekwC+Dfx+hXG+1h+LxUXrR/v8KxiAKP3B7du6HvwNPAM8D6zzNDVvWAccAZqAVwGl7RehP5NJN2Uy6VfD4YhI/Wiff0s8WtovUr/dAJ4Bjq9Kat5yEAgDF7W4KP3hcES0fpY/f5sBiNRfmgI8DozqI596+kn6+g7QsWM79fX1nmb5v7K4uEQyeYfXxl7n1y/+ptKQAeDM8t8V9T/9hc/R9/FedtSq/ttJxsZe58Vfnao0ZCAWi/9H/WvWNtIYDhEKBQkEqpntB49Siny+QCadZz6VqTRkADizPAWoqH9oaKi6SXrE8PBwpfAAcCbQ19vfAIwDW0t7uro7+drXB9mzZ7dHKVaXq1f/zs9+OsytiX/Yw0mgVP6U6e/u7uIb3/wKe3o+5GGW1ePqu9f4yXPPMzFxyx5OAl2xWBw0/fX1QZpbIkQiDeze7TwGAoEASikCNeIKszMprl0dJ5VasIeTQNeyAZTp98uJr6MZQRLoCm7f1vUE8EX7nh/88Lu+OfkB2tpa2dGxnVde+bM93AxcB/ag6f/Rj7/Hnh6f6e/s4OULZVV/M3A9HI449K9rjdLYGKKnp4dAIFDzWzjSyNrmNSRv33Xob4lHHfr3799fjY9h1bl8+bL932bgenD7tq7vUzwJAPj80SM88sinUEr5alvfup7792dIJMbtb4ICdtv1H33qCfo/8+lVz/eD3lpb1zM1NUPivUSZ/nA4UqZ/bXOY5uYm35z8pS0SaQRg+v5smf6WeLRMv1+//aFobJoJqBDwUXukr+8AhULB08S8QKkCH9n7YS6cf9keflgf94lPHkQpfYHcDyj2PrSX83+6YA869DdFG+np6fEuLQ/ZvHkD44myaaBDvzAeDgGb7JGNmzb60gAKhQLr1q/Xw+1olwPb2zf50gCUUrS1tephh/76+mDNzO3/W6JrmvSQQ78w2kNAoz0SDNZhWdYq5VM9CoUCdXWOA9vR3RUKhXxrgHV1dXrYob+urq7SOF/gRr8wGkJ6xI8HPxR1udHmx29/4MFawEr42QAMToQZwMongHT9pUUzgwwcBuDH8h9KJ8DK2qRXAIFAwFQAghBWAayszc/6lVpZmzEAWQirAGQbgBttxgBkYSqACuP8iNs1ALMIKAtRFYAbbX5eA8CsARg0RFUAbubAftZfcKHfVACyEFYBCDcAswZg0BBlAAUX2qQbgKkAZCHKACzhi4CmD8CgI8oA3DQC+dkAzBTAoCPKANxok24AZgogC1EGIL0PwHQCGnQEGYAyFYDLRqBgMOhBRob/B0QZgPQKwEwBDDqiDEB6BaDMIqBBQ5QBFKQ3Apk1AIOGKAOwxF8GNGsAhnJEGYD0TkA3VwHMGoAsRBmA9E5A0whk0BFlAOIrAHM/AIOGKAMQXwG4nAKYNQA5yDIA4RWAmQIYdOQYgFLCfwzkrhHKTAFkIcYAlFLCbwji/sEgZgogB1EGYCoAMwUwlCPLAIRXAMYADDqiDEB2J6AylwENDkQZgOw+AGUuAxociDIA2X0AZgpgcCLLAIRXAG5+DmymALIQZQDirwKYuwIbNEQZgPQ+ANMIZNARZQDSLwOaOwIZdEQZgPTLgOaOQAYdMQ8HVcrdN6Bf9RenAKYPwFCOsApg5ZPbz48HN48GM+iIMgDRjUAuKyBjALIQNQUQ/VwA5b4T0BiAHERVAJI7AYsGaKYAhnJEVQBuSmC/rgG4fTy4qQBkIcYALMvdPfH8qj+ft1zdFhyKVYBBBiEgBawtBXK5nC9/DZbJZCotAmaBDDb9mUyWhoZ6L1PzhHQ6U2l659BvWZZvK4BcLq+HHPqFkQ0Bk9jegKWlNJFIePVSqgJKKRYWlkins/quSbQDIJVKEY+3eJle1VFKMZeaZz41r+9y6F9cWPKtAdx7/54ecugXxmQI+BuwqxS5PzVL++aG1UupCswvLJK38synFvRdVwCFTf+N8Vs8tC/mZXpVZ3pmDitvMXnnrr7LoX9i4rZvpwBvvHFFDzn0+5lTp07poSsh4BzweCkyeecea9Y2EY1GvMytKigFi0tLZDNZMpkcM9Mpfcg5igfAA/1vv/UOW7e207ah1ctUq4JSirm5eZYW08zMzJJ474Y+xKH/r29fJZG4wa5dO71Mtepcv55g9PQ5PezQPzw8zNDQkJepecbU1JQeOhfo6+1vAMaBraVoY7iBLVs31KwJFAoFcjmLdDqDUopsNsf9qTly2Zx9WBLopngAlOlvaYnxsd59tLW1UmtfhkqBVbDIZnLMzy9QUIq52RTvvnON1FzZFCAJdMdicYf+zs4Ovjz4Jbp3dnqcfXUYH7/JL174JTdv3rKHk0B3R2ebQ38kEuHYsWMeZ1ldTpw4QSaTsYeSQHcAoK+3fwA4rb+oORYlEg0TCoVq8kTI5/MsLaSZm3WU/gCHxy5dHIV/r79r5w42btpANNpUc2Vxcd1jkbv/fJ8biYlKQx7of+zRgYr6Dx85RO/B/WzZ0l5zC8OWZZFMTnLptTc5/duzlYYcfun86CjA4OBgRf3xeJyjR49WOdPqcvLkSaanpyvtOjwyMjL64Kju6+1/FviWZ5mtLs+NXbp43B6Qrv+xRwdE6X/p/GiZ/sHBQVH6R0ZGjgPYl3u/A/x8dfLxlBcoatUx+o1+cfrtBmABQxQXRBIeJ+UFCWAA+CpFrTpGv9EvTr+jExD4HfDH5cGHgH3AFqDWumNyFBc6/gKcBUYpNn6shNFv9IvR/y/ppbP4R1txSgAAAABJRU5ErkJggg==".fromBase64()
+    ).toBMP32()
+}
